ui code moved to presentation package bottom sheet package moved into dashboard screen package
