--- app/src/main/java/com/ably/tracking/demo/subscriber/Routes.kt
@@ -1,20 +1,0 @@
-package com.ably.tracking.demo.subscriber
-
-sealed interface Routes {
-
-    object Login : Routes {
-        const val path = "login/"
-    }
-
-    object CreateOrder : Routes {
-        const val path = "order/create/"
-    }
-
-    object Dashboard : Routes {
-        const val path = "dashboard/"
-    }
-
-    object OrderArrived : Routes {
-        const val path = "order/arrived"
-    }
-}
