+++ app/src/main/java/com/ably/tracking/demo/subscriber/di/KoinModule.kt
@@ -10,17 +10,17 @@ import com.ably.tracking.demo.subscriber.api.buildOkHttpClient
 import com.ably.tracking.demo.subscriber.api.buildRetrofit
 import com.ably.tracking.demo.subscriber.common.FusedLocationSource
 import com.ably.tracking.demo.subscriber.domain.OrderManager
+import com.ably.tracking.demo.subscriber.presentation.navigation.Navigator
+import com.ably.tracking.demo.subscriber.presentation.screen.createorder.CreateOrderViewModel
+import com.ably.tracking.demo.subscriber.presentation.screen.dashboard.DashboardViewModel
+import com.ably.tracking.demo.subscriber.presentation.screen.login.LoginViewModel
+import com.ably.tracking.demo.subscriber.presentation.screen.orderarrived.OrderArrivedViewModel
 import com.ably.tracking.demo.subscriber.secrets.AndroidBase64Encoder
 import com.ably.tracking.demo.subscriber.secrets.Base64Encoder
 import com.ably.tracking.demo.subscriber.secrets.InMemorySecretsManager
 import com.ably.tracking.demo.subscriber.secrets.SecretsManager
 import com.ably.tracking.demo.subscriber.secrets.SecretsStorage
 import com.ably.tracking.demo.subscriber.secrets.SharedPreferencesSecretsStorage
-import com.ably.tracking.demo.subscriber.ui.screen.Navigator
-import com.ably.tracking.demo.subscriber.ui.screen.createorder.CreateOrderViewModel
-import com.ably.tracking.demo.subscriber.ui.screen.dashboard.DashboardViewModel
-import com.ably.tracking.demo.subscriber.ui.screen.login.LoginViewModel
-import com.ably.tracking.demo.subscriber.ui.screen.orderarrived.OrderArrivedViewModel
 import com.ably.tracking.ui.animation.CoreLocationAnimator
 import com.ably.tracking.ui.animation.LocationAnimator
 import org.koin.androidx.viewmodel.dsl.viewModel
