+++ app/src/main/java/com/ably/tracking/demo/subscriber/presentation/screen/dashboard/DashboardScreenStatePreview.kt
@@ -1,0 +1,12 @@
+package com.ably.tracking.demo.subscriber.presentation.screen.dashboard
+
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import com.ably.tracking.demo.subscriber.domain.OrderState
+
+class DashboardScreenStatePreview : PreviewParameterProvider<DashboardScreenState> {
+    override val values = sequenceOf(
+        DashboardScreenState(
+            orderState = OrderState(OrderState.State.Failed, "Error")
+        )
+    )
+}
