+++ app/src/main/java/com/ably/tracking/demo/subscriber/presentation/screen/orderarrived/OrderArrivedViewModel.kt
@@ -1,0 +1,11 @@
+package com.ably.tracking.demo.subscriber.presentation.screen.orderarrived
+
+import androidx.lifecycle.ViewModel
+import com.ably.tracking.demo.subscriber.presentation.navigation.Navigator
+
+class OrderArrivedViewModel(private val navigator: Navigator) : ViewModel() {
+
+    fun onCreateNewOrderClick() {
+        navigator.navigateToRecreateOrder()
+    }
+}
