+++ app/src/main/java/com/ably/tracking/demo/subscriber/presentation/theme/Shape.kt
@@ -1,0 +1,11 @@
+package com.ably.tracking.demo.subscriber.presentation.theme
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Shapes
+import androidx.compose.ui.unit.dp
+
+val Shapes = Shapes(
+    small = RoundedCornerShape(4.dp),
+    medium = RoundedCornerShape(4.dp),
+    large = RoundedCornerShape(0.dp)
+)
