+++ app/src/main/java/com/ably/tracking/demo/subscriber/presentation/theme/Type.kt
@@ -1,0 +1,15 @@
+package com.ably.tracking.demo.subscriber.presentation.theme
+
+import androidx.compose.material.Typography
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.sp
+
+val Typography = Typography(
+    body1 = TextStyle(
+        fontFamily = FontFamily.Default,
+        fontWeight = FontWeight.Normal,
+        fontSize = 16.sp
+    )
+)
