--- app/src/main/java/com/ably/tracking/demo/subscriber/ui/bottomsheet/LocationUpdateBottomSheetData.kt
@@ -1,12 +1,0 @@
-package com.ably.tracking.demo.subscriber.ui.bottomsheet
-
-import com.ably.tracking.LocationUpdate
-import com.ably.tracking.Resolution
-
-data class LocationUpdateBottomSheetData(
-    val locationUpdate: LocationUpdate?,
-    val resolution: Resolution?,
-    val remainingDistance: Double?,
-    val lastLocationUpdateInterval: Long?,
-    val averageLocationUpdateInterval: Long?
-)
