--- app/src/main/java/com/ably/tracking/demo/subscriber/ui/screen/dashboard/DashboardScreenState.kt
@@ -1,17 +1,0 @@
-package com.ably.tracking.demo.subscriber.ui.screen.dashboard
-
-import com.ably.tracking.LocationUpdate
-import com.ably.tracking.Resolution
-import com.ably.tracking.demo.subscriber.domain.OrderState
-
-data class DashboardScreenState(
-    val orderId: String = "",
-    val orderState: OrderState? = null,
-    val lastLocationUpdateInterval: Long? = null,
-    val averageLocationUpdateInterval: Long? = null,
-    val resolution: Resolution? = null,
-    val remainingDistance: Double? = null,
-    val orderLocation: LocationUpdate? = null,
-    val animatedLocation: LocationUpdate? = null,
-    val showSubscriptionFailedDialog: Boolean = false
-)
