--- app/src/main/java/com/ably/tracking/demo/subscriber/ui/screen/orderarrived/OrderArrivedViewModel.kt
@@ -1,11 +1,0 @@
-package com.ably.tracking.demo.subscriber.ui.screen.orderarrived
-
-import androidx.lifecycle.ViewModel
-import com.ably.tracking.demo.subscriber.ui.screen.Navigator
-
-class OrderArrivedViewModel(private val navigator: Navigator) : ViewModel() {
-
-    fun onCreateNewOrderClick() {
-        navigator.navigateToRecreateOrder()
-    }
-}
