--- app/src/main/java/com/ably/tracking/demo/subscriber/ui/theme/Color.kt
@@ -1,7 +1,0 @@
-package com.ably.tracking.demo.subscriber.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Primary = Color(0xFF03020D)
-val PrimaryVariant = Color(0xFF000000)
-val Accent = Color(0xFFDC0C0C)
