--- app/src/main/java/com/ably/tracking/demo/subscriber/ui/theme/Theme.kt
@@ -1,23 +1,0 @@
-package com.ably.tracking.demo.subscriber.ui.theme
-
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.darkColors
-import androidx.compose.runtime.Composable
-
-private val DarkColorPalette = darkColors(
-    primary = Primary,
-    primaryVariant = PrimaryVariant,
-    secondary = Accent
-)
-
-@Composable
-fun AATSubscriberDemoTheme(
-    content: @Composable () -> Unit
-) {
-    MaterialTheme(
-        colors = DarkColorPalette,
-        typography = Typography,
-        shapes = Shapes,
-        content = content
-    )
-}
